Fix egs_cbct compilation and example
Fix a compiler error that occurred with newer compiler settings due to
ambiguous declaration of isinf and isnan. Also change the z-rotation
parameter in the example input file to be 0 so that it calculates the
first projection instead of a later projection. This avoids a common
user error when trying to produce the blank scan for the first time.